requirements.txt
# REST API
fastapi>=0.68.0
uvicorn>=0.15.0

# gRPC
grpcio>=1.41.0
grpcio-tools>=1.41.0
protobuf>=3.19.0

# Kafka
confluent-kafka>=1.8.2

# Machine Learning
scikit-learn==1.5.2
numpy>=1.21.0
pandas>=1.3.0

# Utilidades
python-dotenv>=0.19.0
requests>=2.26.0 